staticfiles/staticfiles/requirements.2006c4407817.txt
aniso8601==7.0.0
asgiref==3.4.1
boto3==1.18.2
botocore==1.21.2
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-heroku==0.3.1
django-storages==1.11.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
jmespath==0.10.0
Pillow==8.3.1
promise==2.3
psycopg2==2.9.1
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
Rx==1.6.1
s3transfer==0.5.0
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
whitenoise==5.3.0
